move nh code to original solver2d
